feat: add resize for Sidebar
